updates con inversa.

+++ src/matrix_functions.h
@@ -8,7 +8,7 @@ void multiply (size_t rows1, size_t columns1, size_t rows2, size_t columns2 ,int
 void concat (size_t rows, size_t columns1,size_t columns2, int m1[rows][columns1], int m2[rows][columns2], int answer[rows][columns1+columns2]);
 void printMatrix(size_t rows, size_t columns, int m[rows][columns]);
 void printMatrixFloat(size_t rows, size_t columns, float m[rows][columns]);
-void inverse(size_t size, int m[size][size], float answer[size][size]);
+void inverse(size_t size, int m[size][size], int answer[size][size]);
 void cofactor(int size, int m[size][size], int ans[size][size], int forbidden_row, int forbidden_col);
 int determinant(size_t size, int m[size][size]);
 void adjoint(int size, int m[size][size],int answer[size][size]);
